Finish new digital object form functional component conversion, plus other related fixes
