+++ app/javascript/hyacinth_ui_v1/components/digital_objects/DigitalObjects.js
@@ -2,7 +2,8 @@ import React from 'react';
 import { Route, Switch } from 'react-router-dom';
 
 import PageNotFound from '../layout/PageNotFound';
-import DigitalObjectNew from './DigitalObjectNew';
+import DigitalObjectNew from './new/DigitalObjectNew';
+import DigitalObjectNewForm from './new/DigitalObjectNewForm';
 import DigitalObjectSearch from './DigitalObjectSearch';
 import DigitalObject from './DigitalObject';
 
@@ -11,6 +12,7 @@ function DigitalObjects() {
     <div>
       <Switch>
         <Route exact path="/digital_objects" component={DigitalObjectSearch} />
+        <Route path="/digital_objects/new/:projectStringKey/:digitalObjectType" component={DigitalObjectNewForm} />
         <Route path="/digital_objects/new" component={DigitalObjectNew} />
         <Route path="/digital_objects/:id" component={DigitalObject} />
         <Route component={PageNotFound} />
